drm/i915: Eliminate magic numbers "0" and "1" from color plane
According to many computer science sources - magic values
in code _are_ _bad_. For many reasons: the reason is that "0"
or "1" or whatever magic values confuses and doesn't give any
info why this parameter is this value and what it's meaning
is.
I renamed "0" to COLOR_PLANE_Y and "1" to COLOR_PLANE_UV,
because we in fact already use this naming in many other places
and function names, when dealing with color planes.

Signed-off-by: Stanislav Lisovskiy <stanislav.lisovskiy@intel.com>